+++ gherkin/java/src/test/java/gherkin/ParserTest.java
@@ -3,7 +3,11 @@ package gherkin;
 import gherkin.ast.GherkinDocument;
 import gherkin.deps.com.google.gson.Gson;
 import gherkin.deps.com.google.gson.JsonParser;
+import gherkin.pickles.Compiler;
+import gherkin.pickles.Pickle;
 import org.junit.Test;
+
+import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
@@ -85,6 +89,68 @@ public class ParserTest {
                 "\"comments\":[]," +
                 "\"type\":\"GherkinDocument\"}"),
                 jsonParser.parse(gson.toJson(gherkinDocument)));
+    }
+
+    @Test
+    public void doc_string_content_type() throws Exception {
+        Gson gson = new Gson();
+        JsonParser jsonParser = new JsonParser();
+        String in1 = "" +
+                "Feature: Foo\n" +
+                "  Scenario: Bar\n" +
+                "    Given x\n" +
+                "      \"\"\"json\n" +
+                "      {}\n" +
+                "      \"\"\"";
+        TokenMatcher matcher = new TokenMatcher();
+        Parser<GherkinDocument> parser = new Parser<>(new AstBuilder());
+
+        GherkinDocument gherkinDocument =  parser.parse(in1, matcher);
+
+        assertEquals(
+                jsonParser.parse("" +
+                        "{\"feature\":{\"tags\":[]," +
+                        "    \"language\":\"en\"," +
+                        "    \"keyword\":\"Feature\"," +
+                        "    \"name\":\"Foo\"," +
+                        "    \"children\":[{" +
+                        "        \"tags\":[]," +
+                        "        \"keyword\":\"Scenario\"," +
+                        "        \"name\":\"Bar\"," +
+                        "        \"steps\":[{" +
+                        "            \"keyword\":\"Given \"," +
+                        "            \"text\":\"x\"," +
+                        "            \"argument\":{" +
+                        "                \"content\":\"{}\"," +
+                        "                \"type\":\"DocString\"," +
+                        "                \"location\":{\"line\":4,\"column\":7}," +
+                        "                \"contentType\":\"json\"}," +
+                        "            \"type\":\"Step\"," +
+                        "            \"location\":{\"line\":3,\"column\":5}}]," +
+                        "        \"type\":\"Scenario\"," +
+                        "        \"location\":{\"line\":2,\"column\":3}}]," +
+                        "    \"type\":\"Feature\"," +
+                        "    \"location\":{\"line\":1,\"column\":1}}," +
+                        "\"comments\":[]," +
+                        "\"type\":\"GherkinDocument\"}"),
+                jsonParser.parse(gson.toJson(gherkinDocument)));
+
+
+        Compiler compiler = new Compiler();
+        List<Pickle> pickles = compiler.compile(gherkinDocument);
+
+        assertEquals( jsonParser.parse("" +
+                        "[{\"name\":\"Bar\",\"language\":\"en\"," +
+                        "  \"steps\":[{" +
+                        "        \"text\":\"x\"," +
+                        "        \"arguments\": [{" +
+                        "            \"location\":{\"line\":4,\"column\":7}," +
+                        "            \"content\":\"{}\",\"contentType\":\"json\"}]," +
+                        "        \"locations\":[{\"line\":3,\"column\":11}]}]," +
+                        "  \"tags\":[],\"locations\":[{\"line\":2,\"column\":3}]}]" ),
+                jsonParser.parse(gson.toJson(pickles))
+                );
+
     }
 
 }